// src/lib/strapi/types/api/evenement.ts
// Interface automatically generated by schemas-to-ts

import { Media } from '../common/Media';
import { Media_Plain } from '../common/Media';
import { AdminPanelRelationPropertyModification } from '../common/AdminPanelRelationPropertyModification';

export interface Evenement {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    description?: any;
    type: string;
    image_principale: { data: Media };
    bouton_texte?: string;
    bouton_lien?: string;
    rank?: number;
    date_fin: Date;
  };
}
export interface Evenement_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  description?: any;
  type: string;
  image_principale: Media_Plain;
  bouton_texte?: string;
  bouton_lien?: string;
  rank?: number;
  date_fin: Date;
}

export interface Evenement_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  description?: any;
  type: string;
  image_principale: number;
  bouton_texte?: string;
  bouton_lien?: string;
  rank?: number;
  date_fin: Date;
}

export interface Evenement_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  description?: any;
  type: string;
  image_principale: AdminPanelRelationPropertyModification<Media_Plain>;
  bouton_texte?: string;
  bouton_lien?: string;
  rank?: number;
  date_fin: Date;
}